moved auto_cky_lib back to server. new bug where auto_ckt_sim is None when called
