+++ Example/Shared/example_shared/__init__.py
@@ -181,3 +181,4 @@ auto_ckt_sim = Rpc(
     return_type=AutoCktOutput,
     docstring="Simulation on the Server",
 )
+
